+++ components/Info.vue
@@ -46,9 +46,12 @@ export default {
     margin-top 3rem
     padding 2rem
     border-radius 1rem
-    background var(--whiteColor)
-    box-shadow 0px 0px 8px var(--shadowColor)
+    background $whiteColor
+    box-shadow 0px 0px 8px $shadowColor
     transition all .5s ease-in-out
+    @media (prefers-color-scheme: dark)
+      background $whiteDarkColor
+      box-shadow 0px 0px 8px $shadowDarkColor
 .no-bg
   padding-top ($headerHeight + 2rem)
   .info-content