+++ layouts/404.vue
@@ -32,16 +32,20 @@ export default {
                 0px 0px 6px rgba(0,0,0,.2)
   .error-description
     font-size 2rem
-    color var(--textColor)
+    color $textColor
     font-weight 400
     opacity .5
     transition color .5s ease-in-out
+    @media (prefers-color-scheme: dark)
+      color $textDarkColor
   .error-home
     display inline-block
     margin 2rem 0 4rem
     font-size 2rem
-    color var(--accentColor)
+    color $accentColor
     transition color .5s ease-in-out
+    @media (prefers-color-scheme: dark)
+      color $accentDarkColor
     &:hover
       opacity .8
 </style>