+++ layouts/GlobalLayout.vue
@@ -38,9 +38,7 @@ export default {
   },
   data() {
     return {
-      colorScheme: {
-        hasNativeSupport: false
-      }
+      colorScheme: {}
     }
   },
   computed: {
@@ -58,18 +56,26 @@ export default {
     }
   },
   mounted() {
-    this.colorScheme = prefersColorScheme()
+    const localTheme = window.localStorage.getItem('defaultTheme') || ''
+    const defaultTheme = localTheme || this.$themeConfig.defaultTheme
+    this.colorScheme = prefersColorScheme(defaultTheme)
   },
   methods: {
     changeScheme() {
       this.colorScheme.scheme =
         this.colorScheme.scheme === 'light' ? 'dark' : 'light'
+      window.localStorage.setItem('defaultTheme', this.colorScheme.scheme)
     }
   }
 }
 </script>
 
 <style lang="stylus" scoped>
+#global-layout
+  background $bgColor
+  transition all .5s ease-in-out
+  @media (prefers-color-scheme: dark)
+    background $bgDarkColor
 .switch
   position absolute
   right 1rem
@@ -81,15 +87,19 @@ export default {
   display inline-block
   width 52px
   height 31px
-  border 1px solid var(--borderColor)
+  border 1px solid $borderColor
   border-radius 20px
-  box-shadow var(--borderColor) 0 0 0 0 inset
-  background-color var(--bgColor)
+  box-shadow $borderColor 0 0 0 0 inset
+  background-color $bgColor
   background-clip content-box
   -webkit-appearance none
   user-select none
   outline none
   transition all .5s ease-in-out
+  @media (prefers-color-scheme: dark)
+    border 1px solid $borderDarkColor
+    box-shadow $borderDarkColor 0 0 0 0 inset
+    background-color $bgDarkColor
 .switch-input:before
   content ''
   position absolute
@@ -98,15 +108,18 @@ export default {
   width 29px
   height 29px
   border-radius 20px
-  box-shadow 0 1px 3px var(--shadowColor)
-  background-color var(--whiteColor)
+  box-shadow 0 1px 3px $shadowColor
+  background-color $whiteColor
   transition all 0.5s ease-in-out
-.switch-input:checked {
-  background-color var(--bgColor)
+  @media (prefers-color-scheme: dark)
+    box-shadow 0 1px 3px $shadowDarkColor
+    background-color $whiteDarkColor
+.switch-input:checked
+  background-color $bgColor
   transition background-color .5s
-}
-.switch-input:checked:before {
+  @media (prefers-color-scheme: dark)
+    background-color $bgDarkColor
+.switch-input:checked:before
   left: 21px
   transition left 0.5s ease-in-out
-}
 </style>